requirements.txt
numpy==1.20.2
streamlit==0.80.0
pillow == 8.2.0
keras == 2.4.3
tensorflow==2.3.0